socketioxide

Module adapter

Adapters are responsible for managing the internal state of the server (rooms, sockets, etc…). When a socket joins or leaves a room, the adapter is responsible for updating the state. The default adapter is the LocalAdapter, which stores the state in memory. Other adapters can be made to share the state between multiple servers.

Structs

BroadcastOptions

Options that can be used to modify the behavior of the broadcast methods.

LocalAdapter

The default adapter. Store the state in memory.

Enums

BroadcastFlags

Flags that can be used to modify the behavior of the broadcast methods.

Traits

Adapter

An adapter is responsible for managing the state of the server. This adapter can be implemented to share the state between multiple servers. The default adapter is the LocalAdapter, which stores the state in memory.

Type Aliases

Room

A room identifier