socketioxide::handler::connect

Trait FromConnectParts

pub trait FromConnectParts<A: Adapter>: Sized {
    type Error: Error + Send + 'static;

    // Required method
    fn from_connect_parts(
        s: &Arc<Socket<A>>,
        auth: &Option<Value>,
    ) -> Result<Self, Self::Error>;
}

A trait used to extract the arguments from the connect event. The Result associated type is used to return an error if the extraction fails, in this case the ConnectHandler is not called.

See the connect module doc for more details on connect handler.

See the extract module doc for more details on available extractors.

Required Associated Types

type Error: Error + Send + 'static

The error type returned by the extractor

Required Methods

fn from_connect_parts( s: &Arc<Socket<A>>, auth: &Option<Value>, ) -> Result<Self, Self::Error>

Extract the arguments from the connect event. If it fails, the handler is not called

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.