socketioxide::handler::disconnect

Trait DisconnectHandler

pub trait DisconnectHandler<A: Adapter, T>:
    Send
    + Sync
    + 'static {
    // Required method
    fn call(&self, s: Arc<Socket<A>>, reason: DisconnectReason);
}

Define a handler for the disconnect event. It is implemented for closures with up to 16 arguments. They must implement the FromDisconnectParts trait.

See the disconnect module doc for more details on disconnect handler.

See the extract module doc for more details on available extractors.

Required Methods

fn call(&self, s: Arc<Socket<A>>, reason: DisconnectReason)

Call the handler with the given arguments.