socketioxide::handler::disconnect

Trait FromDisconnectParts

pub trait FromDisconnectParts<A: Adapter>: Sized {
    type Error: Error + 'static;

    // Required method
    fn from_disconnect_parts(
        s: &Arc<Socket<A>>,
        reason: DisconnectReason,
    ) -> Result<Self, Self::Error>;
}

A trait used to extract the arguments from the disconnect event. The Result associated type is used to return an error if the extraction fails, in this case the DisconnectHandler is not called.

See the disconnect module doc for more details on disconnect handler.

See the extract module doc for more details on available extractors.

Required Associated Types

type Error: Error + 'static

The error type returned by the extractor

Required Methods

fn from_disconnect_parts( s: &Arc<Socket<A>>, reason: DisconnectReason, ) -> Result<Self, Self::Error>

Extract the arguments from the disconnect event. If it fails, the handler is not called

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.