socketioxide::handler::message

Trait MessageHandler

pub trait MessageHandler<A: Adapter, T>:
    Send
    + Sync
    + 'static {
    // Required method
    fn call(&self, s: Arc<Socket<A>>, v: Value, ack_id: Option<i64>);
}

Define a handler for the connect event. It is implemented for closures with up to 16 arguments. They must implement the FromMessageParts trait or the FromMessage trait for the last one.

See the message module doc for more details on message handler.

See the extract module doc for more details on available extractors.

Required Methods

fn call(&self, s: Arc<Socket<A>>, v: Value, ack_id: Option<i64>)

Call the handler with the given arguments