socketioxide::operators

Trait RoomParam

pub trait RoomParam: 'static {
    type IntoIter: Iterator<Item = Room>;

    // Required method
    fn into_room_iter(self) -> Self::IntoIter;
}

A trait for types that can be used as a room parameter.

String, Vec<String>, Vec<&str>, &'static str and const arrays are implemented by default.

Required Associated Types

type IntoIter: Iterator<Item = Room>

The type of the iterator returned by into_room_iter.

Required Methods

fn into_room_iter(self) -> Self::IntoIter

Convert self into an iterator of rooms.