socketioxide::socket

Enum DisconnectReason

pub enum DisconnectReason {
    TransportClose,
    MultipleHttpPollingError,
    PacketParsingError,
    TransportError,
    HeartbeatTimeout,
    ClientNSDisconnect,
    ServerNSDisconnect,
    ClosingServer,
}

All the possible reasons for a Socket to be disconnected from a namespace.

It can be used as an extractor in the on_disconnect handler.

Variants

TransportClose

The client gracefully closed the connection

MultipleHttpPollingError

The client sent multiple polling requests at the same time (it is forbidden according to the engine.io protocol)

PacketParsingError

The client sent a bad request / the packet could not be parsed correctly

TransportError

The connection was closed (example: the user has lost connection, or the network was changed from WiFi to 4G)

HeartbeatTimeout

The client did not send a PONG packet in the ping timeout delay

ClientNSDisconnect

The client has manually disconnected the socket using socket.disconnect()

ServerNSDisconnect

The socket was forcefully disconnected from the namespace with Socket::disconnect or with SocketIo::delete_ns

ClosingServer

The server is being closed