socketioxide::socket

Struct Socket

pub struct Socket<A: Adapter = LocalAdapter> {
    pub id: Sid,
    /* private fields */
}

A Socket represents a client connected to a namespace. It is used to send and receive messages from the client, join and leave rooms, etc. The socket struct itself should not be used directly, but through a SocketRef.

Fields
id: Sid

The socket id

Implementations

impl<A: Adapter> Socket<A>

pub fn on<H, T>(&self, event: impl Into<Cow<'static, str>>, handler: H)
where H: MessageHandler<A, T>, T: Send + Sync + 'static,

Registers a MessageHandler for the given event.

See the message module doc for more details on message handler.

See the extract module doc for more details on available extractors.

Simple example with a sync closure:

#[derive(Debug, Serialize, Deserialize)]
struct MyData {
    name: String,
    age: u8,
}

let (_, io) = SocketIo::new_svc();
io.ns("/", |socket: SocketRef| {
    // Register a handler for the "test" event and extract the data as a `MyData` struct
    // With the Data extractor, the handler is called only if the data can be deserialized as a `MyData` struct
    // If you want to manage errors yourself you can use the TryData extractor
    socket.on("test", |socket: SocketRef, Data::<MyData>(data)| {
        println!("Received a test message {:?}", data);
        socket.emit("test-test", &MyData { name: "Test".to_string(), age: 8 }).ok(); // Emit a message to the client
    });
});

Example with a closure and an acknowledgement + binary data:

#[derive(Debug, Serialize, Deserialize)]
struct MyData {
    name: String,
    age: u8,
}

let (_, io) = SocketIo::new_svc();
io.ns("/", |socket: SocketRef| {
    // Register an async handler for the "test" event and extract the data as a `MyData` struct
    // Extract the binary payload as a `Vec<Bytes>` with the Bin extractor.
    // It should be the last extractor because it consumes the request
    socket.on("test", |socket: SocketRef, Data::<MyData>(data), ack: AckSender| async move {
        println!("Received a test message {:?}", data);
        tokio::time::sleep(std::time::Duration::from_secs(1)).await;
        ack.send(&data).ok(); // The data received is sent back to the client through the ack
        socket.emit("test-test", &MyData { name: "Test".to_string(), age: 8 }).ok(); // Emit a message to the client
    });
});

pub fn on_disconnect<C, T>(&self, callback: C)
where C: DisconnectHandler<A, T> + Send + Sync + 'static, T: Send + Sync + 'static,

Registers a disconnect handler.

You can register only one disconnect handler per socket. If you register multiple handlers, only the last one will be used.

See the disconnect module doc for more details on disconnect handler.

See the extract module doc for more details on available extractors.

The callback will be called when the socket is disconnected from the server or the client or when the underlying connection crashes. A DisconnectReason is passed to the callback to indicate the reason for the disconnection.

Example

let (_, io) = SocketIo::new_svc();
io.ns("/", |socket: SocketRef| {
    socket.on("test", |socket: SocketRef| async move {
        // Close the current socket
        socket.disconnect().ok();
    });
    socket.on_disconnect(|socket: SocketRef, reason: DisconnectReason| async move {
        println!("Socket {} on ns {} disconnected, reason: {:?}", socket.id, socket.ns(), reason);
    });
});

pub fn emit<T: ?Sized + Serialize>( &self, event: impl AsRef<str>, data: &T, ) -> Result<(), SendError>

Emits a message to the client

If you provide tuple-like data (tuple, arrays), it will be considered as multiple arguments. Therefore if you want to send an array as the first argument of the payload, you need to wrap it in an array or a tuple. Vec will be always considered as a single argument though.

Emitting binary data

To emit binary data, you must use a data type that implements Serialize as binary data. Currently if you use Vec<u8> it will be considered as a number sequence and not binary data. To counter that you must either use a special type like Bytes or use the serde_bytes crate. If you want to emit generic data that may contains binary, use rmpv::Value rather than serde_json::Value otherwise the binary data will also be serialized as a number sequence.

Errors

When encoding the data a SendError::Serialize may be returned.

If the underlying engine.io connection is closed a SendError::Socket(SocketError::Closed) will be returned and the provided data to be send will be given back in the error.

If the packet buffer is full, a SendError::Socket(SocketError::InternalChannelFull) will be returned and the provided data to be send will be given back in the error. See SocketIoBuilder::max_buffer_size option for more infos on internal buffer config

Example

let (_, io) = SocketIo::new_svc();
io.ns("/", |socket: SocketRef| {
    socket.on("test", |socket: SocketRef, Data::<Value>(data)| async move {
        // Emit a test message to the client
        socket.emit("test", &data).ok();

        // Emit a test message with multiple arguments to the client
        socket.emit("test", &("world", "hello", 1)).ok();

        // Emit a test message with an array as the first argument
        let arr = [1, 2, 3, 4];
        socket.emit("test", &[arr]).ok();
    });
});

Binary Example with the bytes crate

let (_, io) = SocketIo::new_svc();
io.ns("/", |socket: SocketRef| {
    socket.on("test", |socket: SocketRef, Data::<(String, Bytes, Bytes)>(data)| async move {
        // Emit a test message to the client
        socket.emit("test", &data).ok();

        // Emit a test message with multiple arguments to the client
        socket.emit("test", &("world", "hello", Bytes::from_static(&[1, 2, 3, 4]))).ok();

        // Emit a test message with an array as the first argument
        let arr = [1, 2, 3, 4];
        socket.emit("test", &[arr]).ok();
    });
});

pub fn emit_with_ack<T: ?Sized + Serialize, V>( &self, event: impl AsRef<str>, data: &T, ) -> Result<AckStream<V>, SendError>

Emits a message to the client and wait for acknowledgement.

See emit() for more details on how to emit data.

The acknowledgement has a timeout specified in the config (5s by default) (see SocketIoBuilder::ack_timeout) or with the timeout() operator.

To get acknowledgements, an AckStream is returned. It can be used in two ways:

As a Stream: It will yield all the ack responses with their corresponding socket id received from the client. It can useful when broadcasting to multiple sockets and therefore expecting more than one acknowledgement. If you want to get the socket from this id, use io::get_socket().

As a Future: It will yield the first ack response received from the client. Useful when expecting only one acknowledgement.

Errors

If the packet encoding failed an EncodeError is immediately returned.

If the socket is full or if it has been closed before receiving the acknowledgement, an SendError::Socket will be immediately returned and the value to send will be given back.

If the client didn’t respond before the timeout, the AckStream will yield an AckError::Timeout. If the data sent by the client is not deserializable as V, an AckError::Decode will be yielded.

Basic example

let (_, io) = SocketIo::new_svc();
io.ns("/", |socket: SocketRef| {
    socket.on("test", |socket: SocketRef, Data::<Value>(data)| async move {
        // Emit a test message and wait for an acknowledgement with the timeout specified in the config
        match socket.emit_with_ack::<_, Value>("test", &data).unwrap().await {
            Ok(ack) => println!("Ack received {:?}", ack),
            Err(err) => println!("Ack error {:?}", err),
        }
   });
});

pub fn join(&self, rooms: impl RoomParam) -> Result<(), A::Error>

Joins the given rooms.

If the room does not exist, it will be created.

Errors

When using a distributed adapter, it can return an Adapter::Error which is mostly related to network errors. For the default LocalAdapter it is always an Infallible error

pub fn leave(&self, rooms: impl RoomParam) -> Result<(), A::Error>

Leaves the given rooms.

If the room does not exist, it will do nothing

Errors

When using a distributed adapter, it can return an Adapter::Error which is mostly related to network errors. For the default LocalAdapter it is always an Infallible error

pub fn leave_all(&self) -> Result<(), A::Error>

Leaves all rooms where the socket is connected.

Errors

When using a distributed adapter, it can return an Adapter::Error which is mostly related to network errors. For the default LocalAdapter it is always an Infallible error

pub fn rooms(&self) -> Result<Vec<Room>, A::Error>

Gets all rooms where the socket is connected.

Errors

When using a distributed adapter, it can return an Adapter::Error which is mostly related to network errors. For the default LocalAdapter it is always an Infallible error

pub fn connected(&self) -> bool

Return true if the socket is connected to the namespace.

A socket is considered connected when it has been successfully handshaked with the server and that all connect middlewares have been executed.

pub fn to(&self, rooms: impl RoomParam) -> BroadcastOperators<A>

Selects all clients in the given rooms except the current socket.

If you want to include the current socket, use the within() operator.

Example

let (_, io) = SocketIo::new_svc();
io.ns("/", |socket: SocketRef| {
    socket.on("test", |socket: SocketRef, Data::<Value>(data)| async move {
        let other_rooms = "room4".to_string();
        // In room1, room2, room3 and room4 except the current
        socket
            .to("room1")
            .to(["room2", "room3"])
            .to(vec![other_rooms])
            .emit("test", &data);
    });
});

pub fn within(&self, rooms: impl RoomParam) -> BroadcastOperators<A>

Selects all clients in the given rooms.

It does include the current socket contrary to the to() operator.

Example

let (_, io) = SocketIo::new_svc();
io.ns("/", |socket: SocketRef| {
    socket.on("test", |socket: SocketRef, Data::<Value>(data)| async move {
        let other_rooms = "room4".to_string();
        // In room1, room2, room3 and room4 including the current socket
        socket
            .within("room1")
            .within(["room2", "room3"])
            .within(vec![other_rooms])
            .emit("test", &data);
    });
});

pub fn except(&self, rooms: impl RoomParam) -> BroadcastOperators<A>

Filters out all clients selected with the previous operators which are in the given rooms.

Example

let (_, io) = SocketIo::new_svc();
io.ns("/", |socket: SocketRef| {
    socket.on("register1", |socket: SocketRef, Data::<Value>(data)| async move {
        socket.join("room1");
    });
    socket.on("register2", |socket: SocketRef, Data::<Value>(data)| async move {
        socket.join("room2");
    });
    socket.on("test", |socket: SocketRef, Data::<Value>(data)| async move {
        // This message will be broadcast to all clients in the Namespace
        // except for ones in room1 and the current socket
        socket.broadcast().except("room1").emit("test", &data);
    });
});

pub fn local(&self) -> BroadcastOperators<A>

Broadcasts to all clients only connected on this node (when using multiple nodes). When using the default in-memory LocalAdapter, this operator is a no-op.

Example

let (_, io) = SocketIo::new_svc();
io.ns("/", |socket: SocketRef| {
    socket.on("test", |socket: SocketRef, Data::<Value>(data)| async move {
        // This message will be broadcast to all clients in this namespace and connected on this node
        socket.local().emit("test", &data);
    });
});

pub fn timeout(&self, timeout: Duration) -> ConfOperators<'_, A>

Sets a custom timeout when sending a message with an acknowledgement.

See SocketIoBuilder::ack_timeout for the default timeout.

See emit_with_ack() for more details on acknowledgements.

Example

let (_, io) = SocketIo::new_svc();
io.ns("/", |socket: SocketRef| {
   socket.on("test", |socket: SocketRef, Data::<Value>(data)| async move {
      // Emit a test message in the room1 and room3 rooms, except for the room2 room with the binary payload received, wait for 5 seconds for an acknowledgement
      socket.to("room1")
            .to("room3")
            .except("room2")
            .timeout(Duration::from_secs(5))
            .emit_with_ack::<_, Value>("message-back", &data)
            .unwrap()
            .for_each(|(sid, ack)| async move {
               match ack {
                   Ok(ack) => println!("Ack received, socket {} {:?}", sid, ack),
                   Err(err) => println!("Ack error, socket {} {:?}", sid, err),
               }
            }).await;
   });
});

pub fn broadcast(&self) -> BroadcastOperators<A>

Broadcasts to all clients without any filtering (except the current socket).

Example

let (_, io) = SocketIo::new_svc();
io.ns("/", |socket: SocketRef| {
    socket.on("test", |socket: SocketRef, Data::<Value>(data)| async move {
        // This message will be broadcast to all clients in this namespace
        socket.broadcast().emit("test", &data);
    });
});

pub fn disconnect(self: Arc<Self>) -> Result<(), DisconnectError>

Disconnects the socket from the current namespace,

It will also call the disconnect handler if it is set.

pub fn ns(&self) -> &str

Gets the current namespace path.

pub fn req_parts(&self) -> &Parts

Gets the request info made by the client to connect

It might be used to retrieve the http::Extensions

pub fn transport_type(&self) -> TransportType

Gets the TransportType used by the client to connect with this Socket

It can also be accessed as an extractor:

let (_, io) = SocketIo::new_svc();
io.ns("/", |socket: SocketRef, transport: TransportType| {
    assert_eq!(socket.transport_type(), transport);
});

pub fn protocol(&self) -> ProtocolVersion

Gets the socket.io ProtocolVersion used by the client to connect with this Socket

It can also be accessed as an extractor:

Example

let (_, io) = SocketIo::new_svc();
io.ns("/", |socket: SocketRef, v: ProtocolVersion| {
    assert_eq!(socket.protocol(), v);
});