socketioxide

Struct SocketIoBuilder

pub struct SocketIoBuilder<A: Adapter = LocalAdapter> { /* private fields */ }

A builder to create a SocketIo instance. It contains everything to configure the socket.io server with a SocketIoConfig. It can be used to build either a Tower Layer or a Service.

Implementations

impl<A: Adapter> SocketIoBuilder<A>

pub fn new() -> Self

Creates a new SocketIoBuilder with default config

pub fn req_path(self, req_path: impl Into<Cow<'static, str>>) -> Self

The path to listen for socket.io requests on.

Defaults to “/socket.io”.

pub fn ping_interval(self, ping_interval: Duration) -> Self

The interval at which the server will send a ping packet to the client.

Defaults to 25 seconds.

pub fn ping_timeout(self, ping_timeout: Duration) -> Self

The amount of time the server will wait for a ping response from the client before closing the connection.

Defaults to 20 seconds.

pub fn max_buffer_size(self, max_buffer_size: usize) -> Self

The maximum number of packets that can be buffered per connection before being emitted to the client. If the buffer if full the emit() method will return an error

Defaults to 128 packets.

pub fn max_payload(self, max_payload: u64) -> Self

The maximum size of a payload in bytes. If a payload is bigger than this value the emit() method will return an error.

Defaults to 100 kb.

pub fn transports<const N: usize>(self, transports: [TransportType; N]) -> Self

Allowed transports on this server

The transports array should have a size of 1 or 2

Defaults to : [TransportType::Polling, TransportType::Websocket]

pub fn ack_timeout(self, ack_timeout: Duration) -> Self

The amount of time the server will wait for an acknowledgement from the client before closing the connection.

Defaults to 5 seconds.

pub fn connect_timeout(self, connect_timeout: Duration) -> Self

The amount of time before disconnecting a client that has not successfully joined a namespace.

Defaults to 45 seconds.

pub fn with_config(self, config: SocketIoConfig) -> Self

Sets a custom SocketIoConfig created previously for this SocketIoBuilder

pub fn with_parser(self, parser: ParserConfig) -> Self

Sets a custom ParserConfig for this SocketIoBuilder

let (io, layer) = SocketIo::builder()
    .with_parser(ParserConfig::msgpack())
    .build_layer();

pub fn with_adapter<B: Adapter>(self) -> SocketIoBuilder<B>

Sets a custom Adapter for this SocketIoBuilder

pub fn with_state<S: Clone + Send + Sync + 'static>(self, state: S) -> Self

Add a custom global state for the SocketIo instance. This state will be accessible from every handler with the State extractor. You can set any number of states as long as they have different types. The state must be cloneable, therefore it is recommended to wrap it in an Arc if you want shared state.

pub fn build_layer(self) -> (SocketIoLayer<A>, SocketIo<A>)

Builds a SocketIoLayer and a SocketIo instance

The layer can be used as a tower layer

pub fn build_svc(self) -> (SocketIoService<NotFoundService, A>, SocketIo<A>)

Builds a SocketIoService and a SocketIo instance

This service will be a standalone service that return a 404 error for every non-socket.io request It can be used as a hyper service

pub fn build_with_inner_svc<S: Clone>( self, svc: S, ) -> (SocketIoService<S, A>, SocketIo<A>)

Builds a SocketIoService and a SocketIo instance with an inner service

It can be used as a hyper service